update query fixed

+++ DCheat_Server/DCheat_Server/messageHandler.py
@@ -173,11 +173,14 @@ class ForkingRequestHandler(socketserver.BaseRequestHandler):
         courseIndex = select_course_index(courseName)
         
         try:
-            delete_ban_list_in_course(courseIndex)
-            delete_allow_list_in_course(courseIndex)
+            if len(banList[0]) is not 0:
+                delete_ban_list_in_course(courseIndex)
+            
+            if len(allowList[0]) is not 0:
+                delete_allow_list_in_course(courseIndex)
             dao.commit()
         except:
-            dao.rollbak()
+            dao.rollback()
             self.request.send("0".encode('utf-8'))
             return
